research/requirements.txt
pillow
ultralytics
streamlit